LayoutTests/imported/w3c/web-platform-tests/domxpath/callback-interface-expected.txt

CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: XPathNSResolver does not have a lookupNamespaceURI method.
CONSOLE MESSAGE: XPathNSResolver does not have a lookupNamespaceURI method.
CONSOLE MESSAGE: XPathNSResolver does not have a lookupNamespaceURI method.

PASS callable resolver 
PASS callable resolver: result is not cached 
PASS callable resolver: abrupt completion from Call 
FAIL callable resolver: no 'lookupNamespaceURI' lookups assert_equals: expected 0 but got 1
PASS object resolver 
PASS object resolver: 'lookupNamespaceURI' is not cached 
FAIL object resolver: abrupt completion from Get assert_throws_dom: function "() => {
    document.evaluate("/foo:bar", document.documentElement, resolver);
  }" threw object "[object Object]" that is not a DOMException NAMESPACE_ERR: property "code" is equal to undefined, expected 14
FAIL object resolver: 'lookupNamespaceURI' is thruthy and not callable assert_true: Timed out waiting for error expected true got false
FAIL object resolver: 'lookupNamespaceURI' is falsy and not callable assert_true: Timed out waiting for error expected true got false
